requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.7.0
boto3==1.34.63
botocore==1.34.63
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cors==1.0.1
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
future==0.18.2
gevent==24.2.1
greenlet==3.0.3
gyp==0.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lumigo-core==0.0.9
lumigo-tracer==1.1.234
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.4
oauthlib==3.2.0
olefile==0.46
openai==1.14.3
pandas==2.0.3
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.0
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymssql==2.2.11
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==2.0.0
s3transfer==0.10.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlparse==0.5.0
systemd-python==234
tldextract==5.1.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.18
usb-creator==0.3.7
virtualenv==20.25.0
wadllib==1.3.6
Werkzeug==3.0.1
xdg==5
xkit==0.0.0
XlsxWriter==3.2.0
zipp==1.0.0
zope.event==5.0
zope.interface==6.2
